Move stats to bin (#1407)
### Description

We have 4 utility commands that show us statistics of different
circuits. Currently, these features were compiled with the tests,
causing dependency and feature flags management troubles. The commands
also cause confusion with the ignored test flags.

We can enjoy many benefits if we keep these commands in binary
executables.

- We can compile them on demand. We remove compilation overheads for the
main components and tests.
- Stats commands are confined in the `bin` directory, so we don't have
confusing testing code.

### Issue Link

It was #1246. After exploring, I found room for improving the status
quo.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Move the underlying logics of `make stats_state_circuit`, `make
stats_evm_circuit`, `make stats_copy_circuit`, and `make
evm_exec_steps_occupancy` to `bin` directory.
- Add a feature flag, `stats` for these specific tools.
- Remove `DisplayTable` and replace all its use with `cli-table`'s
feature. This has the benefit that the line of codes in the codebase
reflects more on the core zkevm functionality. After we used
`cli-table`, the table also rendered differently, see images below.
- Move `cli-table` from `dev-dependency` to `dependency`, but as an
optional dependency.
- Remove dead dependency `subtle` and `criterion`

After:
<img width="508" alt="Screenshot 2023-05-14 at 3 26 43 PM"
src="https://github.com/privacy-scaling-explorations/zkevm-circuits/assets/3391420/a4d32ce1-3559-4057-852e-b9b358f18553">
Before:
<img width="471" alt="Screenshot 2023-05-14 at 3 27 53 PM"
src="https://github.com/privacy-scaling-explorations/zkevm-circuits/assets/3391420/a073763d-71e1-4c15-876b-7e4252970875">


### Rationale

- Some undesirable results were introduced.
- We left the Instrument struct unchanged. The `Instrument` is buried
deep in the ExecutionConfig. We intend to deal with it in future PRs to
keep the diff of this one slim.
- We changed some visibilities of structs and functions in evm circuits
from `pub(crate)` to `pub`

### How Has This Been Tested?

Manual run `make stats_state_circuit`, `make stats_evm_circuit`, `make
stats_copy_circuit`, and `make evm_exec_steps_occupancy` and see if they
work.
