+++ Makefile
@@ -58,15 +58,15 @@ exp_bench: ## Run Exp Circuit benchmarks
 circuit_benches: evm_bench state_bench ## Run All Circuit benchmarks
 
 stats_state_circuit: # Print a table with State Circuit stats by ExecState/opcode
-	@cargo test -p zkevm-circuits --features=test,warn-unimplemented get_state_states_stats -- --nocapture --ignored
+	@cargo run --bin stats --features stats -- state
 
 stats_evm_circuit: # Print a table with EVM Circuit stats by ExecState/opcode
-	@cargo test -p zkevm-circuits --features=test,warn-unimplemented get_evm_states_stats -- --nocapture --ignored
+	@cargo run --bin stats --features stats -- evm
 
 stats_copy_circuit: # Print a table with Copy Circuit stats by ExecState/opcode
-	@cargo test -p zkevm-circuits --features=test,warn-unimplemented get_copy_states_stats -- --nocapture --ignored
+	@cargo run --bin stats --features stats -- copy
 
 evm_exec_steps_occupancy: # Print a table for each EVM-CellManager CellType with the top 10 occupancy ExecutionSteps associated
-	@cargo test -p zkevm-circuits --release get_exec_steps_occupancy --features=test,warn-unimplemented -- --nocapture --ignored
+	@cargo run --bin stats --features stats -- exec
 
 .PHONY: clippy doc fmt test test_benches test-all evm_bench state_bench circuit_benches evm_exec_steps_occupancy stats_state_circuit stats_evm_circuit stats_copy_circuit help
