+++ zkevm-circuits/Cargo.toml
@@ -33,20 +33,18 @@ maingate =  { git = "https://github.com/privacy-scaling-explorations/halo2wrong"
 integer =   { git = "https://github.com/privacy-scaling-explorations/halo2wrong", tag = "v2023_04_20" }
 libsecp256k1 = "0.7"
 num-bigint = { version = "0.4" }
-subtle = "2.4"
 rand_chacha = "0.3"
 snark-verifier = { git = "https://github.com/privacy-scaling-explorations/snark-verifier", tag = "v2023_04_20", default-features = false, features = ["loader_halo2", "system_halo2"] }
+cli-table = { version = "0.4", optional = true }
 
 [dev-dependencies]
 bus-mapping = { path = "../bus-mapping", features = ["test"] }
-criterion = "0.3"
 ctor = "0.1.22"
 ethers-signers = "0.17.0"
 hex = "0.4.3"
 itertools = "0.10.1"
 mock = { path = "../mock" }
 pretty_assertions = "1.0.0"
-cli-table = "0.4"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.78"
 
@@ -55,3 +53,8 @@ default = []
 test = ["ethers-signers", "mock", "bus-mapping/test"]
 test-circuits = []
 warn-unimplemented = ["eth-types/warn-unimplemented"]
+stats = ["warn-unimplemented", "dep:cli-table"]
+
+[[bin]]
+name = "stats"
+required-features = ["stats"]
