+++ zkevm-circuits/src/lib.rs
@@ -31,9 +31,6 @@ pub mod table;
 #[cfg(any(feature = "test", test))]
 pub mod test_util;
 
-#[cfg(any(feature = "test", test))]
-mod stats;
-
 pub mod tx_circuit;
 pub mod util;
 pub mod witness;
